Make the return type of GetPIDTargetFromPacket signed
PID targets are integers representing thousandths of a degree. The
software code has been assuming that these can be negative, but the CAN
library has been decoding them as unsigned ints.

In principle, we could do it using unsigned ints (by ensuring that we
always zero the encoders when all of the motors are at the proper joint
limit) but allowing negative values allows us more flexibility and is
less bug-prone.

Unrelated: Fix header file location for PortTemplate.c. In the software
repo we sometimes actually compile this file for testing purposes.